refactor: split useGap hook into utility + hook
Extract SDK infrastructure (client cache, factory, lazy init) from
hooks/useGap.ts into utilities/gapClient.ts so non-React files no
longer depend on a hooks module. Simplify the useGap hook to a
synchronous useMemo derived from chain ID.
