+++ hooks/useGap.ts
@@ -1,146 +1,31 @@
 "use client";
 
-import type { TNetwork } from "@show-karma/karma-gap-sdk";
-import { GAP } from "@show-karma/karma-gap-sdk/core/class/GAP";
-import { GapIndexerClient } from "@show-karma/karma-gap-sdk/core/class/karma-indexer/GapIndexerClient";
-import { Networks } from "@show-karma/karma-gap-sdk/core/consts";
-import { useCallback, useEffect, useMemo, useState } from "react";
+import { useMemo } from "react";
 import { useAccount } from "wagmi";
-import { envVars } from "@/utilities/enviromentVars";
-import { getGapRpcConfig } from "@/utilities/gapRpcConfig";
-import {
-  appNetwork,
-  gapSupportedNetworks,
-  getChainIdByName,
-  getChainNameById,
-} from "@/utilities/network";
-
-const gapClients: Record<number, GAP> = {};
-
-const isSupportedNetwork = (network: string): network is TNetwork =>
-  Object.hasOwn(Networks, network);
-
-const getSupportedNetworkForChain = (chainID: number): TNetwork | null => {
-  const candidate = getChainNameById(chainID);
-  return isSupportedNetwork(candidate) ? candidate : null;
-};
-
-const findDefaultSupportedChainId = (): number | undefined => {
-  const fallbackChain = appNetwork.find((chain) => isSupportedNetwork(getChainNameById(chain.id)));
-  return fallbackChain?.id;
-};
-
-export const getDefaultGapChainId = (): number | undefined => findDefaultSupportedChainId();
-
-const createGapInstance = (network: TNetwork): GAP => {
-  const apiUrl = envVars.NEXT_PUBLIC_GAP_INDEXER_URL;
-  return new GAP({
-    globalSchemas: false,
-    network,
-    rpcUrls: getGapRpcConfig(),
-    ...(apiUrl?.trim()
-      ? {
-          apiClient: new GapIndexerClient(apiUrl),
-        }
-      : {}),
-  });
-};
-
-let allNetworksInitialized = false;
-
-/**
- * Eagerly initialize GAP instances for every supported network so that the
- * SDK's AllGapSchemas.findSchema() always finds an existing instance with
- * rpcConfig populated instead of creating a bare one without it.
- */
-const ensureAllNetworksInitialized = () => {
-  if (allNetworksInitialized) return;
-  allNetworksInitialized = true;
-
-  for (const network of Object.keys(Networks) as TNetwork[]) {
-    const expectedChainId = Networks[network]?.chainId;
-    const resolvedChainId = getChainIdByName(network);
-    if (expectedChainId && resolvedChainId === expectedChainId && !gapClients[resolvedChainId]) {
-      try {
-        gapClients[resolvedChainId] = createGapInstance(network);
-      } catch (error) {
-        console.error(
-          `GAP::Failed to pre-initialize client for ${network} (chain ${resolvedChainId})`,
-          error
-        );
-      }
-    }
-  }
-};
-
-export const getGapClient = (chainID: number): GAP => {
-  ensureAllNetworksInitialized();
-
-  const network = getSupportedNetworkForChain(chainID);
-  if (!network) {
-    throw new Error(`GAP::Unsupported chain ${chainID}`);
-  }
-  const networkChainId = getChainIdByName(network);
-  return (
-    gapClients[networkChainId] ??
-    (() => {
-      const client = createGapInstance(network);
-      gapClients[networkChainId] = client;
-      return client;
-    })()
-  );
-};
+import { getDefaultGapChainId, getGapClient } from "@/utilities/gapClient";
 
 export const useGap = () => {
-  const [gap, setGapClient] = useState<GAP>();
-  const [isUpdating, setIsUpdating] = useState(false);
   const { chain } = useAccount();
-  const defaultSupportedChainId = useMemo(findDefaultSupportedChainId, []);
 
-  const updateGapClient = useCallback(
-    async (chainId: number) => {
-      if (isUpdating) return;
-      setIsUpdating(true);
+  const gap = useMemo(() => {
+    const defaultChainId = getDefaultGapChainId();
+    const targetChainId = chain?.id ?? defaultChainId;
+    if (!targetChainId) return undefined;
 
-      const network = getSupportedNetworkForChain(chainId);
-
-      if (!network) {
-        // Unsupported chain - switch to first supported network
-        const firstSupportedChain = gapSupportedNetworks[0];
-        console.warn(
-          `GAP::Unsupported chain ${chainId}. Switching to ${firstSupportedChain.name}...`
-        );
-
-        setGapClient(getGapClient(firstSupportedChain.id));
-        setIsUpdating(false);
-        return;
+    try {
+      return getGapClient(targetChainId);
+    } catch {
+      // Unsupported chain — fall back to default
+      if (defaultChainId && defaultChainId !== targetChainId) {
+        try {
+          return getGapClient(defaultChainId);
+        } catch {
+          /* noop */
+        }
       }
-
-      // Supported chain - normal flow
-      const normalizedChainId = getChainIdByName(network);
-
-      try {
-        const client = getGapClient(normalizedChainId);
-        setGapClient(client);
-      } catch (error) {
-        console.error("GAP::Failed to initialize client", error);
-        setGapClient(undefined);
-      } finally {
-        setIsUpdating(false);
-      }
-    },
-    [isUpdating]
-  );
-
-  useEffect(() => {
-    const targetChainId = chain?.id ?? defaultSupportedChainId;
-    if (!targetChainId) {
-      setGapClient(undefined);
-      return;
+      return undefined;
     }
+  }, [chain?.id]);
 
-    updateGapClient(targetChainId);
-  }, [chain?.id, defaultSupportedChainId, updateGapClient]);
-
-  return useMemo(() => ({ gap, updateGapClient }), [gap, updateGapClient]);
+  return { gap };
 };
