+++ utilities/gapClient.ts
@@ -1,0 +1,83 @@
+import type { TNetwork } from "@show-karma/karma-gap-sdk";
+import { GAP } from "@show-karma/karma-gap-sdk/core/class/GAP";
+import { GapIndexerClient } from "@show-karma/karma-gap-sdk/core/class/karma-indexer/GapIndexerClient";
+import { Networks } from "@show-karma/karma-gap-sdk/core/consts";
+import { envVars } from "@/utilities/enviromentVars";
+import { getGapRpcConfig } from "@/utilities/gapRpcConfig";
+import { appNetwork, getChainIdByName, getChainNameById } from "@/utilities/network";
+
+const gapClients: Record<number, GAP> = {};
+
+const isSupportedNetwork = (network: string): network is TNetwork =>
+  Object.hasOwn(Networks, network);
+
+const getSupportedNetworkForChain = (chainID: number): TNetwork | null => {
+  const candidate = getChainNameById(chainID);
+  return isSupportedNetwork(candidate) ? candidate : null;
+};
+
+const findDefaultSupportedChainId = (): number | undefined => {
+  const fallbackChain = appNetwork.find((chain) => isSupportedNetwork(getChainNameById(chain.id)));
+  return fallbackChain?.id;
+};
+
+export const getDefaultGapChainId = (): number | undefined => findDefaultSupportedChainId();
+
+const createGapInstance = (network: TNetwork): GAP => {
+  const apiUrl = envVars.NEXT_PUBLIC_GAP_INDEXER_URL;
+  return new GAP({
+    globalSchemas: false,
+    network,
+    rpcUrls: getGapRpcConfig(),
+    ...(apiUrl?.trim()
+      ? {
+          apiClient: new GapIndexerClient(apiUrl),
+        }
+      : {}),
+  });
+};
+
+let allNetworksInitialized = false;
+
+/**
+ * Eagerly initialize GAP instances for every supported network so that the
+ * SDK's AllGapSchemas.findSchema() always finds an existing instance with
+ * rpcConfig populated instead of creating a bare one without it.
+ */
+const ensureAllNetworksInitialized = () => {
+  if (allNetworksInitialized) return;
+  allNetworksInitialized = true;
+
+  for (const network of Object.keys(Networks) as TNetwork[]) {
+    const expectedChainId = Networks[network]?.chainId;
+    const resolvedChainId = getChainIdByName(network);
+    if (expectedChainId && resolvedChainId === expectedChainId && !gapClients[resolvedChainId]) {
+      try {
+        gapClients[resolvedChainId] = createGapInstance(network);
+      } catch (error) {
+        console.error(
+          `GAP::Failed to pre-initialize client for ${network} (chain ${resolvedChainId})`,
+          error
+        );
+      }
+    }
+  }
+};
+
+export const getGapClient = (chainID: number): GAP => {
+  ensureAllNetworksInitialized();
+
+  const network = getSupportedNetworkForChain(chainID);
+  if (!network) {
+    throw new Error(`GAP::Unsupported chain ${chainID}`);
+  }
+  const networkChainId = getChainIdByName(network);
+  return (
+    gapClients[networkChainId] ??
+    (() => {
+      const client = createGapInstance(network);
+      gapClients[networkChainId] = client;
+      return client;
+    })()
+  );
+};
